+++ showoci_data.py
@@ -471,13 +471,18 @@ class ShowOCIData(object):
             if len(list_ip_sec_connections) > 0:
                 retStr += " + IPSEC (" + str(len(list_ip_sec_connections)) + ")"
 
+            # check if Virtual Circuits
+            list_virtual_circuits = self.service.search_multi_items(self.service.C_NETWORK, self.service.C_NETWORK_VC, 'drg_id', drg_id)
+            if len(list_virtual_circuits) > 0:
+                retStr += " + Fastconnect (" + str(len(list_virtual_circuits)) + ")"
+
             # Check Remote Peering
             rpcs = self.service.search_multi_items(self.service.C_NETWORK, self.service.C_NETWORK_RPC, 'drg_id', drg_id)
             if len(rpcs) > 0:
                 retStr += " + Remote Peering (" + str(len(rpcs)) + ")"
 
             # check transit routing
-            if drg_attachment['route_table_id'] != "None":
+            if drg_attachment['route_table_id'] != "None" and drg_attachment['route_table_id'] != "":
                 route_table = str(self.__get_core_network_route(drg_attachment['route_table_id']))
                 retStr += " + Transit Route(" + route_table + ")"
 
@@ -951,7 +956,29 @@ class ShowOCIData(object):
         data = []
         try:
             drgs = self.service.search_multi_items(self.service.C_NETWORK, self.service.C_NETWORK_DRG, 'region_name', region_name, 'compartment_id', compartment['id'])
-            return drgs
+            for drg in drgs:
+                drg_id = drg['id']
+                val = {'id': drg['id'],
+                       'name': drg['name'],
+                       'time_created': drg['time_created'],
+                       'redundancy': drg['redundancy'],
+                       'compartment_name': drg['compartment_name'],
+                       'compartment_id': drg['compartment_id'],
+                       'defined_tags': drg['defined_tags'],
+                       'freeform_tags': drg['freeform_tags'],
+                       'region_name': drg['region_name'],
+                       'ip_sec_connections': self.service.search_multi_items(self.service.C_NETWORK, self.service.C_NETWORK_IPS, 'drg_id', drg_id),
+                       'virtual_circuits': self.service.search_multi_items(self.service.C_NETWORK, self.service.C_NETWORK_VC, 'drg_id', drg_id),
+                       'remote_peerings': self.service.search_multi_items(self.service.C_NETWORK, self.service.C_NETWORK_RPC, 'drg_id', drg_id),
+                       'vcns': []
+                       }
+
+                # Add VCNs
+                drg_attachments = self.service.search_multi_items(self.service.C_NETWORK, self.service.C_NETWORK_DRG_AT, 'drg_id', drg_id)
+                for da in drg_attachments:
+                    val['vcns'].append(self.service.search_unique_item(self.service.C_NETWORK, self.service.C_NETWORK_VCN, 'id', da['vcn_id']))
+                data.append(val)
+            return data
 
         except Exception as e:
             self.__print_error("__get_core_network_drg", e)
@@ -1563,7 +1590,11 @@ class ShowOCIData(object):
 
                     if 'vnic_details' in vnic:
                         if 'display_name' in vnic['vnic_details']:
-                            val = {'id': vnic['vnic_id'], 'desc': vnic['vnic_details']['display_name'] + str(comp_text), 'details': vnic['vnic_details']}
+                            val = {
+                                'id': vnic['vnic_id'],
+                                'desc': vnic['vnic_details']['display_name'] + str(comp_text),
+                                'details': vnic['vnic_details']
+                            }
                             if 'ip_addresses' in vnic['vnic_details']:
                                 val['ip_addresses'] = vnic['vnic_details']['ip_addresses']
                             vnicdata.append(val)
@@ -2135,6 +2166,7 @@ class ShowOCIData(object):
 
             for dbs in list_autos:
                 freemsg = ",  FreeTier" if dbs['is_free_tier'] else ""
+                freesum = "Free " if dbs['is_free_tier'] else ""
                 value = {'id': str(dbs['id']),
                          'name': str(dbs['db_name']) + " (" + (str(dbs['display_name']) + ") - " + str(dbs['license_model']) + " - " + str(dbs['lifecycle_state']) + " (" + str(dbs['sum_count']) + " OCPUs" + (" AutoScale" if dbs['is_auto_scaling_enabled'] else "") + ") - " + dbs['db_workload'] + " - " + dbs['db_type'] + freemsg),
                          'display_name': dbs['display_name'],
@@ -2148,11 +2180,11 @@ class ShowOCIData(object):
                          'service_console_url': str(dbs['service_console_url']),
                          'time_created': str(dbs['time_created'])[0:16],
                          'connection_strings': str(dbs['connection_strings']),
-                         'sum_info': "Autonomous Database " + str(dbs['db_workload']) + " (OCPUs) - " + dbs['license_model'],
-                         'sum_info_stopped': "Stopped Autonomous Database " + str(dbs['db_workload']) + " (Count) - " + dbs['license_model'],
-                         'sum_info_count': "Autonomous Database " + str(dbs['db_workload']) + " (Count) - " + dbs['license_model'],
+                         'sum_info': "Autonomous Database " + freesum + str(dbs['db_workload']) + " (OCPUs) - " + dbs['license_model'],
+                         'sum_info_stopped': "Stopped Autonomous Database " + freesum + str(dbs['db_workload']) + " (Count) - " + dbs['license_model'],
+                         'sum_info_count': "Autonomous Database " + freesum + str(dbs['db_workload']) + " (Count) - " + dbs['license_model'],
                          'sum_count': str(dbs['sum_count']),
-                         'sum_info_storage': "Autonomous Database (TB)",
+                         'sum_info_storage': "Autonomous Database " + freesum + "(TB)",
                          'sum_size_tb': str(dbs['data_storage_size_in_tbs']), 'backups': self.__get_database_autonomous_backups(dbs['backups']),
                          'whitelisted_ips': dbs['whitelisted_ips'],
                          'is_auto_scaling_enabled': dbs['is_auto_scaling_enabled'],